Remove react-router references Correct window checking 2 (Remove WalletRoutes after fix https://github.com/faceyspacey/extract-css-chunks-webpack-plugin/issues/222)
