+++ packages/social-trader/src/components/wallet-widget/wallet-widget-popover-content.tsx
@@ -5,7 +5,7 @@ import { PopoverContent } from "components/popover/popover-content";
 import { StatisticItemList } from "components/statistic-item-list/statistic-item-list";
 import StatisticItem from "components/statistic-item/statistic-item";
 import { StatisticItemContainerBlock } from "components/statistic-item/statistic-item-container.block";
-import { WALLET_TOTAL_PAGE_ROUTE } from "pages/wallet/wallet.routes";
+import { WALLET_TOTAL_PAGE_ROUTE } from "pages/wallet/wallet.paths";
 import React from "react";
 import { useTranslation } from "react-i18next";
 import { formatCurrencyValue } from "utils/formatter";
