+++ packages/social-trader/src/pages/wallet/components/wallet-tables/wallet-tables-total.tsx
@@ -7,11 +7,11 @@ import { HORIZONTAL_POPOVER_POS } from "components/popover/popover";
 import Tooltip from "components/tooltip/tooltip";
 import { TooltipContent } from "components/tooltip/tooltip-content";
 import { WalletData } from "gv-api-web";
+import { WALLET_TOTAL_PAGE_ROUTE } from "pages/wallet/wallet.paths";
 import React from "react";
 import { useTranslation } from "react-i18next";
 
 import useHashTab from "../../services/hashTab.hook";
-import { WALLET_TOTAL_PAGE_ROUTE } from "../../wallet.routes";
 import AllDepositsWithdrawalsRow from "./wallet-deposits-withdrawals/all-deposits-withdrawals-row";
 import WalletDepositsWithdrawals from "./wallet-deposits-withdrawals/wallet-deposits-withdrawals";
 import { WALLET_TOTAL_DEPOSITS_WITHDRAWALS_COLUMNS } from "./wallet-deposits-withdrawals/wallet-deposits-withdrawals.constants";
