+++ packages/social-trader/src/reducers/root-reducer.ts
@@ -1,4 +1,3 @@
-// import { connectRouter } from "connected-react-router";
 import {
   globalSearchReducer,
   GlobalSearchState
@@ -77,7 +76,6 @@ export const sharedRootReducers = {
   fundDetails: fundDetailsReducer,
   followDetails: followDetailsReducer,
   programDetails: programDetailsReducer,
-  // router: connectRouter(history),
   platformData: platformReducer,
   followsData: followsReducer,
   programsData: programsReducer,
@@ -95,9 +93,7 @@ export const sharedRootReducers = {
   followNotifications: followNotificationsReducer,
   programNotifications: programNotificationsReducer,
   fundNotifications: fundNotificationsReducer,
-  // manager: managerReducer,
   wallet: walletReducer,
-  // copyTradingAccounts: CopyTradingAccountsReducer, TODO
   accountSettings: accountSettingsReducer,
   ui: uiReducer
 };
@@ -124,10 +120,8 @@ export type RootState = Readonly<{
   programNotifications: ProgramNotificationsState;
   fundNotifications: FundNotificationsState;
   authData: AuthState;
-  // // router: RouterState;
   alertMessages: AlertMessagesState;
   accountSettings: AccountSettingsState;
   wallet: WalletState;
-  // copyTradingAccounts: CopyTradingAccountsState;
   ui: IUiState;
 }>;
